refactor(sync): extract backend account-join into peers/account_join.rs
